Move copy.digestingReader and tests to separate files
The concept is independent enough from the rest of the
copy implementation, and copy/copy.go is large enough
that splitting it up a bit should improve orientation.

Only moves the code without changes, should not change behavior.
